requirements.txt
transformers
accelerate
datasets
evaluate
peft
bitsandbytes>=0.46.1
pillow
scikit-learn
pandas
numpy
matplotlib
tqdm
sentencepiece
safetensors
opencv-python
pyyaml